Make isAlive() threadsafe for all session types
