+++ src/java.base/share/classes/jdk/internal/foreign/ConfinedSession.java
@@ -25,12 +25,6 @@
 
 package jdk.internal.foreign;
 
-import java.lang.invoke.MethodHandles;
-import java.lang.invoke.VarHandle;
-import java.lang.ref.Cleaner;
-
-import jdk.internal.vm.annotation.ForceInline;
-
 /**
  * A confined session, which features an owner thread. The liveness check features an additional
  * confinement check - that is, calling any operation on this session from a thread other than the
@@ -39,55 +33,8 @@ import jdk.internal.vm.annotation.ForceInline;
  */
 final class ConfinedSession extends MemorySessionImpl {
 
-    private int asyncReleaseCount = 0;
-
-    static final VarHandle ASYNC_RELEASE_COUNT;
-
-    static {
-        try {
-            ASYNC_RELEASE_COUNT = MethodHandles.lookup().findVarHandle(ConfinedSession.class, "asyncReleaseCount", int.class);
-        } catch (Throwable ex) {
-            throw new ExceptionInInitializerError(ex);
-        }
-    }
-
     public ConfinedSession(Thread owner) {
         super(owner, new ConfinedResourceList());
-    }
-
-    @Override
-    @ForceInline
-    public void acquire0() {
-        checkValidState();
-        if (state == MAX_FORKS) {
-            throw tooManyAcquires();
-        }
-        state++;
-    }
-
-    @Override
-    @ForceInline
-    public void release0() {
-        if (Thread.currentThread() == owner) {
-            state--;
-        } else {
-            // It is possible to end up here in two cases: this session was kept alive by some other confined session
-            // which is implicitly released (in which case the release call comes from the cleaner thread). Or,
-            // this session might be kept alive by a shared session, which means the release call can come from any
-            // thread.
-            ASYNC_RELEASE_COUNT.getAndAdd(this, 1);
-        }
-    }
-
-    void justClose() {
-        checkValidState();
-        int asyncCount = (int)ASYNC_RELEASE_COUNT.getVolatile(this);
-        if ((state == 0 && asyncCount == 0)
-                || ((state - asyncCount) == 0)) {
-            state = CLOSED;
-        } else {
-            throw alreadyAcquired(state - asyncCount);
-        }
     }
 
     /**
